# optimize + benchmark muloverflow