Adds a benchmarks and simplifies/improves `MulOverflow` a bit

```
name                          old time/op    new time/op    delta
MulOverflow/single/uint256-6    25.1ns ± 3%    23.0ns ± 1%  -8.22%  (p=0.016 n=5+4)
MulOverflow/single/big-6        77.0ns ±13%    75.7ns ± 4%    ~     (p=0.937 n=5+5)

name                          old alloc/op   new alloc/op   delta
MulOverflow/single/uint256-6     0.00B          0.00B         ~     (all equal)
MulOverflow/single/big-6         0.00B          0.00B         ~     (all equal)

name                          old allocs/op  new allocs/op  delta
MulOverflow/single/uint256-6      0.00           0.00         ~     (all equal)
MulOverflow/single/big-6          0.00           0.00         ~     (all equal)
```