@holiman 2020-06-05
> Actually, we might need to be careful here, I'm not sure this is correct if the same backing slice is used for input.. is it?

Nah that's fine, the `res` from umul is crisp and clean